+++ images/risclet/ui/index.ts
@@ -48,6 +48,7 @@ interface RiscboxRuntime {
 
 interface RiscboxOptions {
     readonly debugTiming?: boolean;
+    readonly guestClockSkew?: number;
     readonly p9Servers: ReadonlyMap<string, {
         connect(): {
             request(bytes: Uint8Array, replyCapacity: number): Promise<